fix problem merge: mix between Alexis's cards and Maxime's structure is good
