+++ gitvs/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import './App.css'
-import CompareTwoProfiles from './Components/CompareTwoProfiles'
-import './Components/Profile1Phone.css'
-import './Components/ProfilesDesktop.css'
+import CompareTwoProfiles from './components/CompareTwoProfiles'
+import './components/Profile1Phone.css'
+import './components/ProfilesDesktop.css'
 
 
 function App() {
