+++ gitvs/src/Components/ButtonMask.js
@@ -1,7 +1,6 @@
-function ButtonMask(id, id2)
-{
-    document.getElementById(id).style.display = 'none';
-    document.getElementById(id2).style.display = 'block';
+function ButtonMask(id, id2) {
+  document.getElementById(id).style.display = 'none'
+  document.getElementById(id2).style.display = 'block'
 }
 
-export default ButtonMask;
+export default ButtonMask
