+++ gitvs/src/components/ButtonMask.js
@@ -1,0 +1,7 @@
+function ButtonMask(id, id2)
+{
+    document.getElementById(id).style.display = 'none';
+    document.getElementById(id2).style.display = 'block';
+}
+
+export default ButtonMask;
