+++ gitvs/src/components/code get profile
@@ -1,0 +1,17 @@
+<GetProfile
+          id="pepita"
+          name="Alox"
+          avatarUrl="https://avatars3.githubusercontent.com/u/8034782?s=460&v=4"
+          publicRepo="6"
+          location="Reims"
+          weapons="JavaScript"
+        />
+
+        <GetProfile
+          id="pepito"
+          name="Maxime-Thibaud"
+          avatarUrl="https://avatars3.githubusercontent.com/u/54673312?v=4"
+          publicRepo="4"
+          location="Reims"
+          weapons="Java"
+    />
