+++ gitvs/src/components/getProfile.css
@@ -10,7 +10,7 @@
     box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
     transition: 0.3s;
     border-radius: 5px;
-    width: 30%;
+    width: 50%;
     height: 65%;
     background-image: url('https://zupimages.net/up/19/43/1e55.png');
     background-size: 100%;
